Se integra capa de componentes "Shared" en Core para uso general de autorizacion de modulos y acciones

+++ src/Core/VRM_Plugin.Core.Abstractions/Data/DTOs/ModuleActionDto.cs
@@ -14,7 +14,7 @@ public class ModuleActionDto
     /// <summary>
     /// ID de la acción en BD (auto-generado por IDENTITY)
     /// </summary>
-    public int ActionKeyId { get; set; }
+    public long ActionKeyId { get; set; }
 
     /// <summary>
     /// ID del componente al que pertenece esta acción (FK)
